+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Support for `dbVersion` field added to Autonomous Database back resource
 - Support for patch and patch history in `database_vm_cluster`
 - Support resource discovery for `monitoring` resources
+- Support resource discovery for `identity` resources
 
 ## 3.79.0 (June 03, 2020)